freeing memory when possible

+++ src/scheduler/FunctionCallServer.cpp
@@ -25,8 +25,8 @@ void FunctionCallServer::stop()
 void FunctionCallServer::doRecv(faabric::transport::Message& header,
                                 faabric::transport::Message& body)
 {
-    assert(header.size() == sizeof(int));
-    int call = static_cast<int>(*header.data());
+    assert(header.size() == sizeof(uint8_t));
+    uint8_t call = static_cast<uint8_t>(*header.data());
     switch (call) {
         case faabric::scheduler::FunctionCalls::MpiMessage:
             this->recvMpiMessage(body);
